Appendix?
Folks, about this reference to an appendix, since it doesn't exist, is it that it just has not yet been done? Or might this be referring to what you may have planned to cover in part7-productionsetupoutsidetheswarm/pipeline-setup-source-control-and-container-registry.md. That is empty, for now (as are a few other pages, but I'll assume you know about them and again just have not yet written them).

Anyway, about this proposed change, I'll understand if you prefer to decline it if you do plan to add the appendix at some point. (or maybe you could phrase it as, "an appendix to be added later", if you wanted to save people looking for it now while reading).

Totally your call. I just wanted to point this out (and mention the other blank pages) if it may be helpful.